move TimeForCall classes to timeForCall package
