+++ src/main/java/by/exadel/internship/mapper/TimeForCallUserMapper.java
@@ -1,9 +1,7 @@
 package by.exadel.internship.mapper;
 
-import by.exadel.internship.dto.TimeForCallDTO;
-import by.exadel.internship.dto.TimeForCallUserDTO;
-import by.exadel.internship.dto.TimeForCallWithUserDTO;
-import by.exadel.internship.entity.TimeForCall;
+import by.exadel.internship.dto.timeForCall.TimeForCallUserDTO;
+import by.exadel.internship.dto.timeForCall.TimeForCallWithUserDTO;
 import by.exadel.internship.entity.TimeForCallUser;
 import org.mapstruct.IterableMapping;
 import org.mapstruct.Mapper;
