+++ src/main/java/by/exadel/internship/service/SchedulingService.java
@@ -1,7 +1,6 @@
 package by.exadel.internship.service;
 
-import by.exadel.internship.dto.TimeForCallUserDTO;
-import by.exadel.internship.dto.TimeForCallWithUserIdDTO;
+import by.exadel.internship.dto.timeForCall.TimeForCallWithUserIdDTO;
 import by.exadel.internship.dto.UserDTO;
 
 import java.util.List;
