+++ src/main/java/by/exadel/internship/service/TimeForCallUserServise.java
@@ -1,12 +1,9 @@
 package by.exadel.internship.service;
 
 
-import by.exadel.internship.dto.TimeForCallUserDTO;
-import by.exadel.internship.dto.TimeForCallWithUserIdDTO;
+import by.exadel.internship.dto.timeForCall.TimeForCallWithUserIdDTO;
 import by.exadel.internship.dto.UserDTO;
-import by.exadel.internship.entity.TimeForCallUser;
 
-import java.util.List;
 import java.util.UUID;
 
 public interface TimeForCallUserServise {
