+++ src/main/java/by/exadel/internship/service/impl/SchedulingServiceImpl.java
@@ -2,13 +2,12 @@ package by.exadel.internship.service.impl;
 
 import by.exadel.internship.dto.InterviewDTO;
 
-import by.exadel.internship.dto.TimeForCallWithUserIdDTO;
+import by.exadel.internship.dto.timeForCall.TimeForCallWithUserIdDTO;
 import by.exadel.internship.dto.UserDTO;
 import by.exadel.internship.dto.enums.FormStatus;
 import by.exadel.internship.dto.enums.UserRole;
 import by.exadel.internship.dto.form.FormFullDTO;
 import by.exadel.internship.exception_handing.InappropriateRoleException;
-import by.exadel.internship.exception_handing.NotFoundException;
 import by.exadel.internship.service.*;
 import by.exadel.internship.util.MDCLog;
 import lombok.RequiredArgsConstructor;
